wrap_ad works with Tensor types, move to router.py
